fix(database): resolve decimal type mismatch and improve connection handling
- Fix type mismatch between PostgreSQL DECIMAL and Rust types
  - Replace f64 with rust_decimal::Decimal for price fields
  - Add rust_decimal and rust_decimal_macros dependencies
  - Update SQLx features to include rust_decimal support

- Update product entity and DTOs
  - Change price field type from f64 to Decimal
  - Update validation logic to use dec! macro
  - Ensure type consistency across the application

- Improve database connection handling
  - Add proper error logging
  - Add connection pool configuration
  - Add connection verification

Technical details:
- Use exact decimal arithmetic for monetary values
- Prevent floating-point precision issues
- Maintain type safety with PostgreSQL
- Fix products list endpoint functionality

Dependencies added:
- rust_decimal = 1.32
- rust_decimal_macros = 1.32

+++ src/domain/entities/product.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 use uuid::Uuid;
@@ -8,7 +9,7 @@ pub struct Product {
     pub id: Uuid,
     pub name: String,
     pub description: String,
-    pub price: f64,
+    pub price: Decimal,
     pub stock: i32,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
@@ -18,7 +19,7 @@ pub struct Product {
 pub struct CreateProductDto {
     pub name: String,
     pub description: String,
-    pub price: f64,
+    pub price: Decimal,
     pub stock: i32,
 }
